requirements.txt
chainlit
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
python-dotenv
chromadb
openai
